# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "超级定价系统 API",
    "description": "\n    ## 宏观错误定价套利引擎\n\n    ### 核心工作区\n    - 💰 **定价研究**: CAPM / Fama-French 三因子 / DCF 估值 / Gap Analysis / 同行对比\n    - 🛰️ **上帝视角 (GodEye)**: 宏观因子引擎 · 证据质量 · 政策雷达 · 结构性衰败 · 跨市场总览\n    - 📂 **研究工作台**: 研究任务持久化 · 状态流转 · 深链重开 · 剧本联动\n    - 🧪 **Quant Lab**: 参数优化 · 风险归因 · 估值历史 · 告警编排 · 数据质量诊断\n\n    ### 私有系统支撑能力\n    - 📊 **内部跨市场复盘**: 模板推荐 · 组合验证 · 执行诊断\n    - 🔗 **另类数据**: 供应链 · 治理 · 人事 · 政策源 · 实体统一\n    - 🔌 **共享运行时支撑**: 行情诊断、历史快照兼容与本地回归脚本\n    - ⚡ **高性能后端**: 异步处理、缓存、诊断与健康检查\n\n    ### API版本\n    - **当前版本**: v4.2.0\n    - **API版本**: v1\n    - **最后更新**: 2026-04-22\n\n    ### 认证\n    研究与只读分析接口默认允许开发态匿名访问；认证、基础设施和管理类接口支持\n    Bearer JWT、Refresh Token 与 X-API-Key。生产环境必须配置 `AUTH_SECRET`，\n    CORS 仅放行显式配置的 `FRONTEND_URL` / `CORS_ORIGINS`。\n\n    ### 限制\n    - 请求频率: 100次/分钟\n    - 数据范围: 最多5年历史数据\n    - 并发实验: 最多10个\n    ",
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "4.2.0"
  },
  "paths": {
...
    "/research-workbench/tasks": {
      "get": {
        "tags": [
          "Research Workbench"
        ],
        "summary": "获取研究工作台任务",
        "operationId": "list_research_tasks_research_workbench_tasks_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "name": "view",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "View"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Research Workbench"
        ],
        "summary": "创建研究工作台任务",
        "operationId": "create_research_task_research_workbench_tasks_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchTaskCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/research-workbench/tasks/bulk-update": {
      "post": {
        "tags": [
          "Research Workbench"
        ],
        "summary": "批量更新研究工作台任务",
        "operationId": "bulk_update_research_tasks_research_workbench_tasks_bulk_update_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchTaskBulkUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/research-workbench/tasks/from-screener": {
      "post": {
        "tags": [
          "Research Workbench"
        ],
        "summary": "从定价筛选器候选创建研究任务",
        "operationId": "create_research_tasks_from_screener_research_workbench_tasks_from_screener_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchTaskFromScreenerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/research-workbench/tasks/{task_id}": {
      "get": {
        "tags": [
          "Research Workbench"
        ],
        "summary": "获取研究工作台任务详情",
        "operationId": "get_research_task_research_workbench_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Research Workbench"
        ],
        "summary": "更新研究工作台任务",
        "operationId": "update_research_task_research_workbench_tasks__task_id__put",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchTaskUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Research Workbench"
        ],
        "summary": "删除研究工作台任务",
        "operationId": "delete_research_task_research_workbench_tasks__task_id__delete",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/research-workbench/tasks/{task_id}/timeline": {
      "get": {
        "tags": [
          "Research Workbench"
        ],
        "summary": "获取研究任务时间线",
        "operationId": "get_research_task_timeline_research_workbench_tasks__task_id__timeline_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/research-workbench/tasks/{task_id}/comments": {
      "post": {
        "tags": [
          "Research Workbench"
        ],
        "summary": "为研究任务添加评论",
        "operationId": "add_research_task_comment_research_workbench_tasks__task_id__comments_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchTaskCommentCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/research-workbench/tasks/{task_id}/comments/{comment_id}": {
      "delete": {
        "tags": [
          "Research Workbench"
        ],
        "summary": "删除研究任务评论",
        "operationId": "delete_research_task_comment_research_workbench_tasks__task_id__comments__comment_id__delete",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "comment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Comment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/research-workbench/tasks/{task_id}/snapshot": {
      "post": {
        "tags": [
          "Research Workbench"
        ],
        "summary": "追加研究任务快照",
        "operationId": "add_research_task_snapshot_research_workbench_tasks__task_id__snapshot_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchTaskSnapshotCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/research-workbench/board/reorder": {
      "post": {
        "tags": [
          "Research Workbench"
        ],
        "summary": "批量更新研究工作台看板顺序",
        "operationId": "reorder_research_board_research_workbench_board_reorder_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchWorkbenchReorderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/research-workbench/briefing/distribution": {
      "get": {
        "tags": [
          "Research Workbench"
        ],
        "summary": "获取每日简报分发配置",
        "operationId": "get_research_briefing_distribution_research_workbench_briefing_distribution_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Research Workbench"
        ],
        "summary": "保存每日简报分发配置",
        "operationId": "update_research_briefing_distribution_research_workbench_briefing_distribution_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchBriefingDistributionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/research-workbench/briefing/dry-run": {
      "post": {
        "tags": [
          "Research Workbench"
        ],
        "summary": "记录每日简报 dry-run 分发",
        "operationId": "run_research_briefing_dry_run_research_workbench_briefing_dry_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchBriefingDryRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/research-workbench/briefing/send": {
      "post": {
        "tags": [
          "Research Workbench"
        ],
        "summary": "发送每日简报到通知通道",
        "operationId": "send_research_briefing_research_workbench_briefing_send_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchBriefingSendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResearchTaskBulkUpdateRequest": {
        "properties": {
          "task_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Task Ids"
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "new",
                  "in_progress",
                  "blocked",
                  "complete",
                  "archived"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "comment": {
            "type": "string",
            "maxLength": 5000,
            "title": "Comment",
            "default": ""
          },
          "author": {
            "type": "string",
            "maxLength": 100,
            "title": "Author",
            "default": "local"
          }
        },
        "type": "object",
        "title": "ResearchTaskBulkUpdateRequest"
      },
      "ResearchTaskCommentCreateRequest": {
        "properties": {
          "author": {
            "type": "string",
            "maxLength": 100,
            "title": "Author",
            "default": "local"
          },
          "body": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 1,
            "title": "Body"
          }
        },
        "type": "object",
        "required": [
          "body"
        ],
        "title": "ResearchTaskCommentCreateRequest"
      },
      "ResearchTaskCreateRequest": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "pricing",
              "cross_market",
              "macro_mispricing",
              "trade_thesis"
            ],
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "string",
            "enum": [
              "new",
              "in_progress",
              "blocked",
              "complete",
              "archived"
            ],
            "title": "Status",
            "default": "new"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": ""
          },
          "symbol": {
            "type": "string",
            "title": "Symbol",
            "default": ""
          },
          "template": {
            "type": "string",
            "title": "Template",
            "default": ""
          },
          "note": {
            "type": "string",
            "title": "Note",
            "default": ""
          },
          "board_order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Board Order"
          },
          "context": {
            "additionalProperties": true,
            "type": "object",
            "title": "Context"
          },
          "snapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResearchTaskSnapshot"
              },
              {
                "type": "null"
              }
            ]
          },
          "refresh_priority_event": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResearchTaskRefreshPriorityEvent"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "type",
          "title"
        ],
        "title": "ResearchTaskCreateRequest"
      },
...
      "ResearchTaskFromScreenerRequest": {
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/ResearchTaskFromScreenerCandidate"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Candidates"
          },
          "source": {
            "type": "string",
            "maxLength": 80,
            "title": "Source",
            "default": "screener"
          }
        },
        "type": "object",
        "required": [
          "candidates"
        ],
        "title": "ResearchTaskFromScreenerRequest"
      },
...
      "ResearchTaskSnapshotCreateRequest": {
        "properties": {
          "snapshot": {
            "$ref": "#/components/schemas/ResearchTaskSnapshot"
          },
          "refresh_priority_event": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResearchTaskRefreshPriorityEvent"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "snapshot"
        ],
        "title": "ResearchTaskSnapshotCreateRequest"
      },
...
    }
  }
}